MOVEONLY: Consensus: Move to consensus/txverify.cpp:
from main.cpp:
-CheckFinalTx
-Consensus::CheckTx
-Consensus::CheckTxInputs
-GetLegacySigOpCount
-GetP2SHSigOpCount
